added experimental support for building ICL with the new C++ compiler clang Since clang is much more pedantic then g++, hundreds of issues had to be fixed. In a future release, a compiler section option will be available through the configure script

+++ ICLQt/src/ThreadedUpdatableTextView.cpp
@@ -46,7 +46,7 @@ namespace icl{
   
   bool ThreadedUpdatableTextView::event ( QEvent * event ){
     ICLASSERT_RETURN_VAL(event,false);
-    switch(event->type()){
+    switch((int)event->type()){
       case ADD_TEXT: 
         append(reinterpret_cast<AddTextEvent*>(event)->text.c_str()); 
         return true;
